backend/service_auth/requirements.txt
fastapi==0.115.0
uvicorn==0.30.3
python-multipart==0.0.20
SQLAlchemy==2.0.32
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv>=0.19.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
cryptography==43.0.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
requests==2.32.3
email-validator==2.1.1
